Refactor namespace SignalR Hub
