+++ src/IoT.Consumer.WebSite/Pages/Components/EventList.razor
@@ -3,7 +3,7 @@
 @using Microsoft.AspNetCore.WebUtilities;
 @using Microsoft.Extensions.Logging
 @using Iot.PnpDashboard.Events
-@using Iot.PnpDashboard.SignalR
+@using Iot.PnpDashboard.EventBroadcast
 
 @implements IAsyncDisposable
 @inject NavigationManager NavigationManager
