+++ src/IoT.Consumer.WebSite/Pages/_Host.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@namespace IoT.Consumer.WebSite.Pages
+@namespace Iot.PnpDashboard.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
     Layout = "_Layout";
