+++ src/IoT.Consumer.WebSite/Pages/_Layout.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
-@namespace IoT.Consumer.WebSite.Pages
+@namespace Iot.PnpDashboard.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
 <!DOCTYPE html>
@@ -10,7 +10,7 @@
     <base href="~/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
     <link href="css/site.css" rel="stylesheet" />
-    <link href="IoT.Consumer.WebSite.styles.css" rel="stylesheet" />
+    <link href="Iot.PnpDashboard.styles.css" rel="stylesheet" />
     <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
 </head>
 <body>
