+++ src/IoT.Consumer.WebSite/Shared/MainLayout.razor
@@ -1,6 +1,6 @@
 ﻿@inherits LayoutComponentBase
 
-@using IoT.Consumer.WebSite.Configuration;
+@using Iot.PnpDashboard.Configuration;
 @inject IAppConfiguration AppConfiguration
